renamed currentWord to currentSecret and added to keyStatuses function
